requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
yt_dlp==2025.1.15
openai-whisper==20240930
langchain_ollama==0.2.2
langchain_core==0.3.31